feat: add presentation slides for introduction, problem, goal, how, paths, and competitors
