+++ web/src/components/svg/Cube.tsx
@@ -1,0 +1,11 @@
+import type { SVGProps } from 'react';
+
+/** Renders a cube icon. */
+const Cube = (props: SVGProps<SVGSVGElement>) => (
+    <svg fill="currentColor" viewBox="-2.4 -2.4 28.8 28.8" {...props}>
+        <path d="M2 18.66l10.5 4.313L23 18.661V6.444L12.5 2.13 2 6.444zm20-.67l-9 3.697V11.102l9-3.68zM12.5 3.213l8.557 3.514-8.557 3.5-8.557-3.5zM3 7.422l9 3.68v10.585L3 17.99z" />
+        <path fill="none" d="M0 0h24v24H0z" />
+    </svg>
+);
+
+export { Cube };
